"Saving before testing angles"

+++ GameEngine/GameEngine/src/Vector2D.cpp
@@ -2,7 +2,7 @@
 
 Vector2D::Vector2D() :x(0), y(0) {}
 
-Vector2D::Vector2D(float x, float y) :x(x), y(y) {}
+Vector2D::Vector2D(int x, int y) :x(x), y(y) {}
 
 //Vector2D&& Vector2D::Add(const Vector2D& vec)
 //{
